Switch to soft tabs.
I was 17 and I made a decision: hard tabs lead to a smaller filesize, so
I'll use that. I am almost 26 now and I will say that I made a mistake.
Ever since I started working with other people, NOBODY seems to like
hard tabs. So for the past few years, I have been hacking around this by
adding editorconfigs to projects I worked on to override this setting.
Part of me was hopeful that the world would come around to this
fantastic realization I had so long ago. But I am grown now and realize
that this dream isn't going to happen and that I need to update my
configs.

Now, to the many months that I'm going to spend fixing this mistake so
I don't have mixed indents anymore.